Fixed some errors.
Fixed c!say command throwing errors when used in DM channel. Also fixed some of the images/videos not sending because of too large file size in c!nudes.

+++ commands/say.js
@@ -2,8 +2,13 @@ module.exports = {
     name: 'say',
     discription: 'Repeats a user message.',
     execute(message, args){
-        const sayMessage = args.join(" ");
-        message.delete().catch(err => console.log(err));
-        message.channel.send(sayMessage);
+        if (message.channel.type === "dm"){
+            const sayMessage = args.join(" ");
+            message.author.send(sayMessage);
+        }else{
+            const sayMessage = args.join(" ");
+            message.delete().catch(err => console.log(err));
+            message.channel.send(sayMessage);
+        }
     }
 }
